chore: folder restructuring
