--- frontend/src/container/MetricsExplorer/Summary/MetricsSearch/index.ts
@@ -1,3 +1,0 @@
-import MetricsSearch from './MetricsSearch';
-
-export default MetricsSearch;
