--- frontend/src/container/MetricsExplorer/Summary/MetricsTable/index.ts
@@ -1,3 +1,0 @@
-import MetricsTable from './MetricsTable';
-
-export default MetricsTable;
