--- frontend/src/container/MetricsExplorer/Summary/MetricsTreemap/index.ts
@@ -1,3 +1,0 @@
-import Treemap from './MetricsTreemap';
-
-export default Treemap;
